style: Seperate components
- MainPage 로딩 로직 변경
- 한 페이지에서 동적 렌더링 사용
- 파일명, 변수명 변경
