+++ src/App.js
@@ -1,20 +1,9 @@
 import './assets/App.css';
-import ActiveSavingsMain from './pages/main/ActiveSavingsMain';
-import NoSavingsMain from './pages/main/NoSavingsMain';
-import MainPage from './pages/main/MainPage';
-import { Routes, Route } from 'react-router-dom'; // BrowserRouter를 추가합니다.
+import MainPage from './pages/MainPage';
 import React from 'react';
 
 function App() {
-    return (
-        <div className="App">
-            <Routes>
-                <Route path="/" element={<MainPage />} />
-                <Route path="/active" element={<ActiveSavingsMain />} />
-                <Route path="/no" element={<NoSavingsMain />} />
-            </Routes>
-        </div>
-    );
+    return <MainPage />;
 }
 
 export default App;
