+++ src/components/main/ActiveSavings.js
@@ -1,0 +1,8 @@
+// 적금이 있을 때
+import SavingsMain from './SavingsMain';
+
+const ActiveSavingsMain = () => {
+    return <SavingsMain containerClass="active" buttonClass="checkCan" />;
+};
+
+export default ActiveSavingsMain;
