+++ src/components/main/NoSavings.js
@@ -1,0 +1,8 @@
+// 적금이 없을 때
+import SavingsMain from './SavingsMain';
+
+const NoSavingsMain = () => {
+    return <SavingsMain containerClass="empty" buttonClass="makeCan" />;
+};
+
+export default NoSavingsMain;
