+++ src/components/main/SavingsMain.js
@@ -1,0 +1,17 @@
+// 공통 컴포넌트
+
+import React from 'react';
+
+import '../../assets/mainPage/MainPage.css';
+import '../../assets/mainPage/ActiveSavingsMain.css';
+import '../../assets/mainPage/NoSavingsMain.css';
+
+const SavingsMain = ({ containerClass, buttonClass }) => {
+    return (
+        <div className={`mobileViewContainer ${containerClass}`}>
+            <button className={`mainBtn ${buttonClass}`}></button>
+        </div>
+    );
+};
+
+export default SavingsMain;
