+++ src/index.js
@@ -2,11 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './assets/index.css';
 import App from './App';
-import { BrowserRouter } from 'react-router-dom';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-    <BrowserRouter>
-        <App />
-    </BrowserRouter>,
-);
+root.render(<App />);
